removed unused code + set random seed
