+++ prepro.py
@@ -1,26 +1,36 @@
 # -*- coding: utf-8 -*-
 # /usr/bin/python3
 '''
-Feb. 2019 by kyubyong park.
-kbpark.linguist@gmail.com.
-https://www.github.com/kyubyong/transformer.
+May 2019 by Kasper Wendel
 
-Preprocess the iwslt 2016 datasets.
+Preprocess the Flickr8k dataset
+Images -> VGG intermediate predictions
+captions -> encoded with Tokenizer SentencePiece as indices
 '''
 
 import errno
 import logging
 import os
 import pickle
+import random as rn
 
 import numpy as np
 import pandas as pd
+from numpy.random import seed
+from tensorflow import set_random_seed
 
-from data_analyse.image import files_to_prediction, load_flickr_set, get_caption_set
-from data_analyse.text import read_captions, clean_captions, train_sp, load_sp, encode_caption_as_ids, add_padding
+from preprocess.dataset import load_flickr_set, get_caption_set
+from preprocess.image import files_to_prediction
+from preprocess.text import read_captions, clean_captions, train_sp, load_sp, encode_caption_as_ids, add_padding
 from utils.hparams import Hparams
 
 logging.basicConfig(level=logging.INFO)
+
+
+def setseed(sd=42):
+    seed(sd)
+    set_random_seed(sd)
+    rn.seed(sd)
 
 
 def prepro(hp):
@@ -122,111 +132,26 @@ def prepro(hp):
         pad_caps = pickle.load(open(pad_path, 'rb'))
     else:
         pad_caps = add_padding(enc_caps, max_length)
-        pickle.dump(pad_caps, file=open(enc_path, "wb"))
+        pickle.dump(pad_caps, file=open(pad_path, "wb"))
 
     # Create the predefined splits with our preprocessed data
     logging.info("# Making data splits with encoded and padded captions")
-    dev = load_flickr_set(images, pad_caps, dev_path, test=False)
-    train = load_flickr_set(images, pad_caps, train_path, test=False)
-    test = load_flickr_set(images, pad_caps, test_path, test=True)
 
-    logging.info("# Writing data pickles")
+    def __write_set(path, name, test):
+        dataset = load_flickr_set(images, pad_caps, path, test=test)
+        p = os.path.join(prepro_path, name)
+        pickle.dump(dataset, file=open(p, 'wb'))
+        logging.info(f"Succesfully saved in {p}")
 
-    def __to_pickle(structure, path):
-        p = os.path.join(prepro_path, path)
-        pickle.dump(structure, file=open(p, 'wb'))
-
-    __to_pickle(dev, 'dev_set.pkl')
-    __to_pickle(train, 'train_set.pkl')
-    __to_pickle(test, 'test_set.pkl')
-
-    # max_length = enc_caps.caption.map(len).max()
-
-    # dev = load_flickr_set(images, enc_caps, dev_path, test=False)
-    # train = load_flickr_set(images, enc_caps, train_path, test=False)
-    # test = load_flickr_set(images, enc_caps, test_path, test=True)
-
-    # # train
-    # _prepro = lambda x: [line.strip() for line in open(x, 'r', encoding='utf-8').read().split("\n") \
-    #                      if not line.startswith("<")]
-    # prepro_train1, prepro_train2 = _prepro(train1), _prepro(train2)
-    # assert len(prepro_train1) == len(prepro_train2), "Check if train source and target files match."
-    #
-    # # eval
-    # _prepro = lambda x: [re.sub("<[^>]+>", "", line).strip() \
-    #                      for line in open(x, 'r', encoding='utf-8').read().split("\n") \
-    #                      if line.startswith("<seg id")]
-    # prepro_eval1, prepro_eval2 = _prepro(eval1), _prepro(eval2)
-    # assert len(prepro_eval1) == len(prepro_eval2), "Check if eval source and target files match."
-    # # test
-    # prepro_test1, prepro_test2 = _prepro(test1), _prepro(test2)
-    # assert len(prepro_test1) == len(prepro_test2), "Check if test source and target files match."
-    #
-    # logging.info("Let's see how preprocessed data look like")
-    # logging.info("prepro_train1:", prepro_train1[0])
-    # logging.info("prepro_train2:", prepro_train2[0])
-    # logging.info("prepro_eval1:", prepro_eval1[0])
-    # logging.info("prepro_eval2:", prepro_eval2[0])
-    # logging.info("prepro_test1:", prepro_test1[0])
-    # logging.info("prepro_test2:", prepro_test2[0])
-    #
-    # logging.info("# write preprocessed files to disk")
-    # os.makedirs("iwslt2016/prepro", exist_ok=True)
-    #
-    # def _write(sents, fname):
-    #     with open(fname, 'w', encoding='utf-8') as fout:
-    #         fout.write("\n".join(sents))
-    #
-    # _write(prepro_train1, "iwslt2016/prepro/train.de")
-    # _write(prepro_train2, "iwslt2016/prepro/train.en")
-    # _write(prepro_train1 + prepro_train2, "iwslt2016/prepro/train")
-    # _write(prepro_eval1, "iwslt2016/prepro/eval.de")
-    # _write(prepro_eval2, "iwslt2016/prepro/eval.en")
-    # _write(prepro_test1, "iwslt2016/prepro/test.de")
-    # _write(prepro_test2, "iwslt2016/prepro/test.en")
-    #
-    # logging.info("# Train a joint BPE model with sentencepiece")
-    # os.makedirs("iwslt2016/segmented", exist_ok=True)
-    # train = '--input=iwslt2016/prepro/train --pad_id=0 --unk_id=1 \
-    #          --bos_id=2 --eos_id=3\
-    #          --model_prefix=iwslt2016/segmented/bpe --vocab_size={} \
-    #          --model_type=bpe'.format(hp.vocab_size)
-    # spm.SentencePieceTrainer.Train(train)
-    #
-    # logging.info("# Load trained bpe model")
-    # sp = spm.SentencePieceProcessor()
-    # sp.Load("iwslt2016/segmented/bpe.model")
-    #
-    # logging.info("# Segment")
-    #
-    # def _segment_and_write(sents, fname):
-    #     with open(fname, "w", encoding='utf-8') as fout:
-    #         for sent in sents:
-    #             pieces = sp.EncodeAsPieces(sent)
-    #             fout.write(" ".join(pieces) + "\n")
-    #
-    # _segment_and_write(prepro_train1, "iwslt2016/segmented/train.de.bpe")
-    # _segment_and_write(prepro_train2, "iwslt2016/segmented/train.en.bpe")
-    # _segment_and_write(prepro_eval1, "iwslt2016/segmented/eval.de.bpe")
-    # _segment_and_write(prepro_eval2, "iwslt2016/segmented/eval.en.bpe")
-    # _segment_and_write(prepro_test1, "iwslt2016/segmented/test.de.bpe")
-    #
-    # logging.info("Let's see how segmented data look like")
-    # print("train1:", open("iwslt2016/segmented/train.de.bpe", 'r', encoding='utf-8').readline())
-    # print("train2:", open("iwslt2016/segmented/train.en.bpe", 'r', encoding='utf-8').readline())
-    # print("eval1:", open("iwslt2016/segmented/eval.de.bpe", 'r', encoding='utf-8').readline())
-    # print("eval2:", open("iwslt2016/segmented/eval.en.bpe", 'r', encoding='utf-8').readline())
-    # print("test1:", open("iwslt2016/segmented/test.de.bpe", 'r', encoding='utf-8').readline())
+    __write_set(dev_path, 'dev_set.pkl', test=False)
+    __write_set(train_path, 'train_set.pkl', test=False)
+    __write_set(test_path, 'test_set.pkl', test=True)
 
 
 if __name__ == '__main__':
     hparams = Hparams()
     parser = hparams.parser
     hp = parser.parse_args()
-    # prepro(hp)
-    dir_path = os.path.join(os.getcwd(), "dataset", "Flickr8k")
-    images_path = os.path.join(dir_path, "Flickr8k_Dataset", "Flicker8k_Dataset")
-    text_path = os.path.join(dir_path, "Flickr8k_text")
-    prepro_path = os.path.join(dir_path, "prepro")
-
-    # logging.info("Done")
+    setseed()
+    prepro(hp)
+    logging.info("Done")
